Refactor summary generation and add background summary trigger
- Moved summary generation logic to a separate API route
- Trigger summary generation in the background via fetch
- Simplified response route by removing detailed summary logging
- Updated thought route to log stream response
- Modified middleware to allow summary API route without authentication
